feat: add overrides & dry-run mode
  - add arxiv, DOI and title overrides to avoid looking for these information if they are known
  - add dry-run option
  - some minor cleaning part

+++ src/pdf_rename/pdf_rename.py
@@ -14,7 +14,6 @@ LICENSE
 
 # System/default
 import sys
-import os
 import re
 import pathlib
 
@@ -22,7 +21,6 @@ import pathlib
 import argparse
 
 # Messaging/logging
-import traceback
 import logging
 from logging.config import dictConfig
 
@@ -32,8 +30,10 @@ import shutil
 # Papers
 import requests
 import pymupdf
-from papers.extract import extract_pdf_metadata
+from papers.extract import pdfhead, _scholar_score
 from papers.encoding import standard_name, family_names
+from papers.config import cached
+from papers import logger
 
 # Bibtex
 import bibtexparser
@@ -127,7 +127,18 @@ def define_argument_parser() -> argparse.ArgumentParser:
     )
 
     # Add overriding options
-    parser.add_argument("-d", "--doi", default=None, type=str, help="Override the DOI")
+    parser.add_argument(
+        "-a",
+        "--arxiv-id",
+        default=None,
+        type=str,
+        help="Assume the paper is an arxiv preprint, extract and then override the DOI using the ID",
+    )
+    parser.add_argument(
+        "-d", "--doi", default=None, type=str, help="Override the DOI (this override has priority over any other ones)"
+    )
+    parser.add_argument("-n", "--dry-run", action="store_true", help="Activate the dry run mode")
+    parser.add_argument("-t", "--title", default=None, type=str, help="Override the title")
 
     # Add arguments
     parser.add_argument("input_pdf", help="The input PDF file to rename")
@@ -169,15 +180,19 @@ def extract_doi_from_pdf(pdf_path: pathlib.Path):
 
 
 def get_metadata_from_crossref(doi) -> tuple[str, str, str, str] | None:
-    """Fetches metadata from CrossRef API using DOI."""
+    """Fetxches metadata from CrossRef API using DOI."""
     url = f"https://api.crossref.org/works/{doi}"
     response = requests.get(url)
 
     if response.status_code == 200:
         data = response.json()
         item = data["message"]
+
         # Extracting relevant metadata
-        year = item["published"]["date-parts"][0][0]  # Year of publication
+        if "published" in item:
+            year = item["published"]["date-parts"][0][0]  # Year of publication
+        else:
+            year = item["date-parts"][0][0]  # Year of publication
         title = item["title"][0]  # The title of the work
         authors = item["author"]
         first_author = authors[0] if authors else None
@@ -189,16 +204,71 @@ def get_metadata_from_crossref(doi) -> tuple[str, str, str, str] | None:
         return None
 
 
-def get_metadata_from_google_scholar(pdf_path: pathlib.Path) -> tuple[str, str, str, str]:
-    # If no information, try to extract some additional part
-    bibtex = extract_pdf_metadata(
-        str(pdf_path.resolve()),
-        search_doi=False,
-        search_fulltext=True,
-        scholar=False,
-        minwords=200,
-        max_query_words=200,
-    )
+@cached('arxiv.json')
+def fetch_bibtex_by_arxiv(arxiv_id: str) -> str:
+    url = f"https://arxiv.org/bibtex/{arxiv_id}"
+    response = requests.get(url)
+    if response.status_code == 200:
+        return response.text
+    else:
+        return f"Error: Unable to fetch BibTeX (HTTP {response.status_code})"
+
+def get_metadata_from_arxiv(arxiv_id: str) -> tuple[str, str, str, str]:
+    bibtex_str = fetch_bibtex_by_arxiv(arxiv_id)
+    bib = bibtexparser.loads(bibtex_str)
+    entry = bib.entries[0]
+
+    # Generate accurate format for author
+    fam_names = family_names(entry.get("author", "unknown").lower())
+    try:
+        fir_names = first_names(entry.get("author", "unknown").lower())
+    except Exception:
+        raise Exception(
+            'The following author entry doesn\'t contain proper author names: "%s"' % (entry.get("author", "unknown"))
+        )
+
+    if (not fam_names) or (fam_names[0] == "unknown"):
+        raise Exception('%s doesn\'t have proper author: "%s"' % (arxiv_id, str(fam_names)))
+
+    return entry["year"], fir_names[0][0].capitalize(), fam_names[0].capitalize(), entry["title"]
+
+
+@cached("scholar-bibtex.json", hashed_key=True)
+def fetch_bibtex_by_fulltext_scholar(txt, assess_results=True):
+    from scholarly import scholarly
+
+    print(txt)
+    # scholarly._get_page = _get_page_fast  # remove waiting time
+    logger.debug(txt)
+    search_query = scholarly.search_pubs(txt)
+
+    # get the most likely match of the first results
+    results = list(search_query)
+    if len(results) > 1 and assess_results:
+        maxscore = 0
+        result = results[0]
+        for res in results:
+            score = _scholar_score(txt, res.bib)
+            if score > maxscore:
+                maxscore = score
+                result = res
+    else:
+        result = results[0]
+
+    return scholarly.bibtex(result)
+
+
+def get_metadata_from_google_scholar(query_txt: str) -> tuple[str, str, str, str]:
+
+    bibtex = fetch_bibtex_by_fulltext_scholar(query_txt)
+    # # If no information, try to extract some additional part
+    # bibtex = extract_txt_metadata(
+    #     query_txt,
+    #     search_doi=False,
+    #     search_fulltext=True,
+    #     scholar=False,
+    #     max_query_words=200,
+    # )
 
     bib = bibtexparser.loads(bibtex)
     entry = bib.entries[0]
@@ -231,22 +301,39 @@ def main():
     output_dir = input_pdf.parent
 
     year, first_initial, last_name, title = (None, None, None, None)
-    doi = args.doi
-    if doi is None:
-        doi = extract_doi_from_pdf(input_pdf)
-    if doi is not None:
-        logger.debug(f"We found a doi: {doi}, let's try crossref")
-        metadata = get_metadata_from_crossref(doi)
-        if metadata is not None:
-            year, first_initial, last_name, title = metadata
-            logger.debug(f"crossref do provide some metadata: {metadata}")
+    if args.arxiv_id is not None:
+        arxiv_id = re.sub(r'v[0-9]*(.pdf)?', '', args.arxiv_id)
+        metadata = get_metadata_from_arxiv(arxiv_id)
+    else:
+        doi = args.doi
+        if doi is None:
+            doi = extract_doi_from_pdf(input_pdf)
+        if doi is not None:
+            logger.debug(f"We found a doi: {doi}, let's try crossref")
+            metadata = get_metadata_from_crossref(doi)
+            if metadata is not None:
+                year, first_initial, last_name, title = metadata
+                logger.debug(f"crossref do provide some metadata: {metadata}")
+            else:
+                logger.warning(f"Metadata couldn't be loaded from DOI: {doi}")
 
-    # We
     if year is None:
         logger.debug(f"Nothing worked up to now, let's try google scholar")
-        metadata = get_metadata_from_google_scholar(input_pdf)
+
+        # Override the title and actually set it as the query
+        if args.title is not None:
+            title = args.title
+            query_text = args.title.lower().strip()
+        else:
+            query_text = pdfhead(str(input_pdf.resolve()), 1, 100, image=False)
+
+        metadata = get_metadata_from_google_scholar(query_text)
         if metadata is not None:
             year, first_initial, last_name, title = metadata
+            if title.lower().strip() != args.title.lower().strip():
+                logger.error(f"The retrieved entry is not the correct one, retrieved title: {title}")
+                sys.exit(-1)
+
             logger.debug(f"google scholar do provide some metadata: {metadata}")
 
     if year is None:
@@ -254,9 +341,11 @@ def main():
         sys.exit(-1)
 
     final_name = "%s - %s. %s - %s.pdf" % (year, first_initial, last_name, title)
-
-    shutil.move(input_pdf, output_dir / final_name)
-    logger.info(f"{input_pdf} renamed to {output_dir}/{final_name}")
+    if not args.dry_run:
+        shutil.move(input_pdf, output_dir / final_name)
+        logger.info(f"{input_pdf} renamed to {output_dir}/{final_name}")
+    else:
+        logger.info(f"[dry-run] {input_pdf} renamed to {output_dir}/{final_name}")
 
 
 ###############################################################################
